requirements.txt
llama-index
faiss-cpu
langchain
pandas 
tqdm 
jsonlines
dotenv
faiss-cpu==1.7.2
streamlit
faiss-cpu
llama-index
PyMuPDF
sentence-transformers
faiss_cpu==1.10.0
fitz==0.0.1.dev2
protobuf==3.20.3
python-dotenv==1.0.1
sentence_transformers==3.4.1
llama-index 
google-generativeai 
python-dotenv